restaurant/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
django-extensions==3.2.1
django-filter==22.1
django_debug_toolbar==3.8.1
localStoragePy==0.2.3
Pillow==9.4.0
sqlparse==0.4.3
